This file was an assignment for week 2
The code asks for an input being number of cheese which has been produced. It then calculates the amount of containers needed to hold all the cheese. After that it calculates the price of those containers and the profit made. It then outputs the number of containers, the cost of the containers, and the profit made from the cheese.